requirements.txt
chardet==5.2.0
darkdetect==0.7.1
greenlet==3.0.3
numpy==2.1.0
pandas==2.2.2
pillow==10.4.0
pyodbc==5.1.0
PyPDF2==3.0.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
pyqtdarktheme==2.1.0
python-dateutil==2.9.0.post0
pytz==2024.1
reportlab==4.2.2
segno==1.6.1
simple_webbrowser==0.2.2
six==1.16.0
SQLAlchemy==2.0.32
typing_extensions==4.12.2
tzdata==2024.1
winotify==1.1.0
